+++ docs/changelog.md
@@ -12,6 +12,7 @@ category: Administration
 ### [Unreleased]
 
 * docs: ensure consistent naming for estimate field
+* feat: add workflow function to start Scrum Poker session (#104)
 
 ### [20.06.1] - 2020-06-19